requirements.txt
requests==2.28.1
pandas==1.4.3
python-telegram-bot==13.13